+++ Ui/Component/Listing/Column/IndexingQueue/AffectedEntities.php
@@ -12,6 +12,7 @@
 namespace Unbxd\ProductFeed\Ui\Component\Listing\Column\IndexingQueue;
 
 use Magento\Ui\Component\Listing\Columns\Column;
+use Magento\Framework\Stdlib\StringUtils;
 use Unbxd\ProductFeed\Model\IndexingQueue;
 use Magento\Framework\UrlInterface;
 use Magento\Framework\Filter\FilterManager;
@@ -24,6 +25,11 @@ use Magento\Framework\View\Element\UiComponent\ContextInterface;
  */
 class AffectedEntities extends Column
 {
+    /**
+     * Max links size for column
+     */
+    const MAX_LINKS = 30;
+
     /**
      * @var IndexingQueue
      */
@@ -41,6 +47,7 @@ class AffectedEntities extends Column
 
     /**
      * AffectedEntities constructor.
+     * @param StringUtils $stringUtils
      * @param IndexingQueue $indexingQueue
      * @param UrlInterface $urlBuilder
      * @param FilterManager $filterManager
@@ -79,7 +86,7 @@ class AffectedEntities extends Column
     {
         if (isset($dataSource['data']['items'])) {
             foreach ($dataSource['data']['items'] as &$item) {
-                $item['affected_entities'] = $this->decorateCell($item['affected_entities']);
+                $item['affected_entities'] = $this->decorateCell($item['affected_entities'], $item['queue_id']);
             }
         }
 
@@ -88,9 +95,10 @@ class AffectedEntities extends Column
 
     /**
      * @param $value
+     * @param $rowId
      * @return string
      */
-    private function decorateCell($value)
+    private function decorateCell($value, $rowId)
     {
         // link on full catalog (by default)
         $cell = '<a href="' . $this->getUrl('catalog/product/index') .'" target="_blank">' . $value .'</a>';
@@ -107,7 +115,28 @@ class AffectedEntities extends Column
             }
 
             $cell = implode(', ', $links);
+            if (count($links) > self::MAX_LINKS) {
+                $cell = $this->formatCell($links, $rowId);
+            }
         }
+
+        return $cell;
+    }
+
+    /**
+     * @param $links
+     * @param $rowId
+     * @return string
+     */
+    private function formatCell(array $links, $rowId)
+    {
+        $viewDetailsUrl = $this->getUrl('unbxd_productfeed/indexing_queue/viewDetails', ['id' => $rowId]);
+
+        $links = array_slice($links, 0, self::MAX_LINKS);
+        $cell = implode(', ', $links);
+
+        $cell .= '...<br/>';
+        $cell .= '<a href="' . $viewDetailsUrl .'" target="_blank">' . __('See Entities Details') . '</a>';
 
         return $cell;
     }